fix OD TV page tests
