+++ src/app/pages/OnDemandTvPage/index.test.jsx
@@ -5,7 +5,7 @@ import { render, act } from '@testing-library/react';
 import { StaticRouter } from 'react-router-dom';
 import { ServiceContextProvider } from '#contexts/ServiceContext';
 import { RequestContextProvider } from '#contexts/RequestContext';
-import OnDemandTvPage from '.';
+import OnDemandTvPage from './OnDemandTvPage';
 import pashtoPageData from '#data/pashto/bbc_pashto_tv/tv_programmes/w13xttn4';
 import * as analyticsUtils from '#lib/analyticsUtils';
 import { ToggleContext } from '#contexts/ToggleContext';
